Use the suffix for scalar types during insertion
As the BE had issues with comparing the values, because of the suffix, this method was replaced with a straighforward approach which represented everything as a string. Instead use this proper method which inserts the literals the way they should be inserted. Old BE might need to be adjusted to work with this change.